+++ bestmobabot/model.py
@@ -1,14 +1,15 @@
+import itertools
 import logging
 import pickle
 from collections import defaultdict
 from operator import itemgetter
-from typing import Any, DefaultDict, Dict, Iterable, List, NamedTuple, Set, Tuple
+from typing import Any, DefaultDict, Dict, Iterable, List, NamedTuple, Optional, Set, Tuple
 
 import numpy
 from pandas import DataFrame, Series
 from scipy import stats
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.model_selection import GridSearchCV, StratifiedKFold, cross_val_score
+from sklearn.model_selection import StratifiedKFold, cross_val_score
 
 from bestmobabot import constants, responses
 from bestmobabot.database import Database
@@ -50,27 +51,22 @@ class Trainer:
         cv = StratifiedKFold(n_splits=self.n_splits, shuffle=True)
 
         self.logger.info('🤖 Adjusting hyper-parameters…')
-        search_cv = GridSearchCV(
+        search_cv = TTestSearchCV(
             estimator,
             param_grid,
             cv=cv,
             scoring=constants.MODEL_SCORING,
-            refit=False,
+            logger=self.logger,
+            alpha=constants.MODEL_SCORING_ALPHA,
         )
-
         search_cv.fit(x, y)
-        estimator.set_params(**search_cv.best_params_)
-
-        # Perform cross-validation.
-        self.logger.info('🤖 Cross validation…')
-        scores: numpy.ndarray = cross_val_score(estimator, x, y, scoring=constants.MODEL_SCORING, cv=cv)
-        score_interval = stats.t.interval(0.95, len(scores) - 1, loc=numpy.mean(scores), scale=stats.sem(scores))
-        self.logger.info(f'🤖 Best score: {search_cv.best_score_:.4f}')
+        score_interval = search_cv.best_confidence_interval_
+        self.logger.info(f'🤖 Best score: {search_cv.best_score_:.4f} ({score_interval[0]:.4f} … {score_interval[1]:.4f})')
         self.logger.info(f'🤖 Best params: {search_cv.best_params_}')
-        self.logger.info(f'🤖 CV score: {scores.mean():.4f} ({score_interval[0]:.4f} … {score_interval[1]:.4f}).')
 
         # Re-train the best model on the entire data.
         self.logger.info('🤖 Refitting…')
+        estimator.set_params(**search_cv.best_params_)
         estimator.fit(x, y)
 
         # Print debugging info.
@@ -112,3 +108,40 @@ class Trainer:
         for hero in heroes:
             for key, value in responses.Hero(hero).feature_dict.items():
                 result[key] += multiplier * value
+
+
+class TTestSearchCV:
+    def __init__(self, estimator, param_grid, *, cv, scoring, logger, alpha=0.95):
+        self.estimator = estimator
+        self.param_grid: Dict[str, Any] = param_grid
+        self.cv = cv
+        self.scoring = scoring
+        self.logger = logger
+        self.alpha = alpha
+        self.best_params_: Optional[Dict[str, Any]] = None
+        self.best_score_: Optional[float] = None
+        self.best_scores_: Optional[numpy.ndarray] = None
+        self.best_confidence_interval_: Optional[numpy.ndarray] = None
+
+    def fit(self, x, y):
+        for values in itertools.product(*self.param_grid.values()):
+            params = dict(zip(self.param_grid.keys(), values))
+            self.estimator.set_params(**params)
+            scores: numpy.ndarray = cross_val_score(self.estimator, x, y, scoring=self.scoring, cv=self.cv)
+            score: float = scores.mean()
+            self.logger.debug(f'🤖 Score: {score:.4f} with {params}.')
+            if not self.is_better_score(score, scores):
+                continue
+            self.logger.debug(f'🤖 Found significantly better score: {score:.4f}.')
+            self.best_params_ = params
+            self.best_score_ = score
+            self.best_scores_ = scores
+            self.best_confidence_interval_ = stats.t.interval(self.alpha, len(scores) - 1, loc=scores.mean(), scale=stats.sem(scores))
+
+    def is_better_score(self, score: float, scores: numpy.ndarray) -> bool:
+        if self.best_params_ is None:
+            return True
+        if score < self.best_score_:
+            return False
+        _, p_value = stats.ttest_ind(self.best_scores_, scores)
+        return p_value < self.alpha