+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Use `3.6.5-wee` docker image
 * Switch to `pipenv`
 * Add `ARENA_EARLY_STOP` parameter
+* Model trainer is now using T-test to compare scores – it should decrease `n_estimators` and improve speed
 * Delete `test` branch
 
 # `v0.4.1`